# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book organizer API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cabinet/": {
            "post": {
                "tags": [
                    "Cabinet"
                ],
                "description": "Create a cabinet entry",
                "operationId": "4b1a94dfbda834c5bf660c952b041126",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "book_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "user_role_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/cabinet/getAllByUser": {
            "get": {
                "tags": [
                    "Cabinet"
                ],
                "description": "Get all books by user",
                "operationId": "0da190292306a9fd68f325a4058a3abf",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "user_role_id",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/cabinet/getCoworkersOfBook": {
            "get": {
                "tags": [
                    "Cabinet"
                ],
                "description": "Get all coworkers of book",
                "operationId": "d9f580afa9f1cb5aece4e984948fed5e",
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "user_role_id",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user": {
                                                        "$ref": "#/components/schemas/User"
                                                    },
                                                    "role": {
                                                        "$ref": "#/components/schemas/UserRole"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "contents": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "UserRole": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "role": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}